Changed custom editor tests to split out reusable helpers
